+++ models/achilles/temp_models/achilles__rawData_406.sql
@@ -9,7 +9,7 @@ MODEL (
 select
   c.condition_concept_id as subject_id,
   p.gender_concept_id,
-  (c.condition_start_year - p.year_of_birth) as count_value
+  (c.condition_start_year - p.year_of_birth)::FLOAT as count_value
 from
   `@src_database`.`@src_schema_omop`.`person` as p
 inner join (
