+++ models/achilles/temp_models/achilles__s_tmpach_0.sql
@@ -13,5 +13,5 @@ select
   cast(getdate() as VARCHAR(255)) as stratum_3,
   cast(null as VARCHAR(255)) as stratum_4,
   cast(null as VARCHAR(255)) as stratum_5,
-  count(distinct person_id) as count_value
+  count(distinct person_id)::FLOAT as count_value
 from `@src_database`.`@src_schema_omop`.`person`
