+++ models/achilles/temp_models/achilles__s_tmpach_102.sql
@@ -10,7 +10,7 @@ with rawData as (
   select
     p1.gender_concept_id as stratum_1,
     year(op1.index_date) - p1.YEAR_OF_BIRTH as stratum_2,
-    count(p1.person_id) as count_value
+    count(p1.person_id)::FLOAT as count_value
   from `@src_database`.`@src_schema_omop`.`person` as p1
   inner join
     (select
