+++ models/achilles/temp_models/achilles__s_tmpach_1020.sql
@@ -10,7 +10,7 @@ with rawData as (
   select
     YEAR(ce.condition_era_start_date) * 100
     + MONTH(ce.condition_era_start_date) as stratum_1,
-    COUNT(ce.person_id) as count_value
+    COUNT(ce.person_id)::FLOAT as count_value
   from
     `@src_database`.`@src_schema_omop`.`condition_era` as ce
   inner join
