+++ models/achilles/temp_models/achilles__s_tmpach_11.sql
@@ -13,7 +13,7 @@ select
   cast(null as VARCHAR(255)) as stratum_3,
   cast(null as VARCHAR(255)) as stratum_4,
   cast(null as VARCHAR(255)) as stratum_5,
-  count(distinct P.person_id) as count_value
+  count(distinct P.person_id)::FLOAT as count_value
 from `@src_database`.`@src_schema_omop`.`person` as P
 where
   not exists
