+++ models/achilles/temp_models/achilles__s_tmpach_1201.sql
@@ -13,7 +13,7 @@ select
   cast(null as VARCHAR(255)) as stratum_3,
   cast(null as VARCHAR(255)) as stratum_4,
   cast(null as VARCHAR(255)) as stratum_5,
-  count(visit_occurrence_id) as count_value
+  count(visit_occurrence_id)::FLOAT as count_value
 from `@src_database`.`@src_schema_omop`.`visit_occurrence` as vo1
 inner join `@src_database`.`@src_schema_omop`.`care_site` as cs1
   on vo1.care_site_id = cs1.care_site_id
