+++ models/achilles/temp_models/achilles__s_tmpach_1425.sql
@@ -13,6 +13,6 @@ select
   cast(null as varchar(255)) as stratum_3,
   cast(null as varchar(255)) as stratum_4,
   cast(null as varchar(255)) as stratum_5,
-  count(*) as count_value
+  count(*)::FLOAT as count_value
 from `@src_database`.`@src_schema_omop`.`payer_plan_period`
 group by payer_source_concept_id
