+++ models/achilles/temp_models/achilles__s_tmpach_2120.sql
@@ -10,7 +10,7 @@ with rawData as (
   select
     YEAR(de.device_exposure_start_date) * 100
     + MONTH(de.device_exposure_start_date) as stratum_1,
-    COUNT(de.person_id) as count_value
+    COUNT(de.person_id)::FLOAT as count_value
   from
     `@src_database`.`@src_schema_omop`.`device_exposure` as de
   inner join
