+++ models/achilles/temp_models/achilles__s_tmpach_2201.sql
@@ -13,7 +13,7 @@ select
   cast(null as VARCHAR(255)) as stratum_3,
   cast(null as VARCHAR(255)) as stratum_4,
   cast(null as VARCHAR(255)) as stratum_5,
-  count(m.PERSON_ID) as count_value
+  count(m.PERSON_ID)::FLOAT as count_value
 from
   `@src_database`.`@src_schema_omop`.`note` as m
 group by m.note_type_CONCEPT_ID
