+++ models/achilles/temp_models/achilles__s_tmpach_300.sql
@@ -13,5 +13,5 @@ select
   cast(null as varchar(255)) as stratum_3,
   cast(null as varchar(255)) as stratum_4,
   cast(null as varchar(255)) as stratum_5,
-  count(distinct provider_id) as count_value
+  count(distinct provider_id)::FLOAT as count_value
 from `@src_database`.`@src_schema_omop`.`provider`
